+++ src/modules/downloader.py
@@ -6,6 +6,8 @@ https://github.com/subzeroid/instagrapi
 """
 import os
 import time
+import json
+import random
 from pathlib import Path
 from urllib3.exceptions import ReadTimeoutError
 from requests.exceptions import ConnectionError as RequestsConnectionError
@@ -19,6 +21,48 @@ class Downloader:
     """
     An Instagram API instance is created by this class and contains a set of all the necessary methods
     to upload content from Instagram to a temporary directory.
+
+    Attributes:
+        :attribute configuration (dict): dictionary with configuration parameters for instagram api communication.
+        :attribute client (object): instance of the instagram api client.
+        :attribute download_methods (dict): dictionary with download methods for instagram api client.
+        :attribute general_settings_list (list): list of general session settings for the instagram api.
+        :attribute device_settings_list (list): list of device settings for the instagram api.
+
+    Methods:
+        :method _get_login_args: get login arguments for the instagram api.
+        :method _create_new_session: create a new session file for the instagram api.
+        :method _handle_relogin: handle re-authentication in the instagram api.
+        :method _load_session: load or create a session.
+        :method _set_session_settings: setting general session settings for the instagram api.
+        :method _validate_session_settings: checking the correctness between the session settings and the configuration settings.
+        :method exceptions_handler: decorator for handling exceptions in the Downloader class.
+        :method login: authentication in instagram api.
+        :method get_post_content: getting the content of a post.
+
+    Examples:
+        >>> configuration = {
+        ...     'username': 'my_username',
+        ...     'password': 'my_password',
+        ...     'session-file': 'data/session.json',
+        ...     'delay-requests': 1,
+        ...     '2fa-enabled': False,
+        ...     '2fa-seed': 'my_seed_secret',
+        ...     'locale': 'en_US',
+        ...     'country-code': '1',
+        ...     'country': 'US',
+        ...     'timezone-offset': 10800,
+        ...     'proxy-dsn': 'http://localhost:8080'
+        ...     'request-timeout': 10,
+        ...     'device-settings': {
+        ...         'app_version': '269.0.0.18.75', 'version_code': '314665256',
+        ...         'manufacturer': 'OnePlus', 'model': '6T Dev', 'device': 'devitron', 'cpu': 'qcom', 'dpi': '480dpi', 'resolution': '1080x1920',
+        ...         'android_release': '8.0.0', 'android_version': '26'
+        ...     }
+        ... }
+        >>> vault = Vault()
+        >>> downloader = Downloader(configuration, vault)
+        >>> status = downloader.get_post_content('shortcode')
     """
     def __init__(
         self,
@@ -38,36 +82,22 @@ class Downloader:
                 :param 2fa-seed (str): seed for two-factor authentication (secret key).
                 :param locale (str): locale for requests.
                 :param country-code (str): country code for requests.
+                :param country (str): country for requests.
                 :param timezone-offset (int): timezone offset for requests.
-                :param user-agent (str): user agent for requests.
                 :param proxy-dsn (str): proxy dsn for requests.
                 :param request-timeout (int): request timeout for requests.
+                :device-settings (dict): dictionary with device settings for requests.
+                    :param app_version (str): application version.
+                    :param version_code (str): version code.
+                    :param manufacturer (str): manufacturer of the device.
+                    :param model (str): model of the device.
+                    :param device (str): device name.
+                    :param cpu (str): cpu name.
+                    :param dpi (str): dpi resolution.
+                    :param resolution (str): screen resolution.
+                    :param android_release (str): android release version.
+                    :param android_version (str): android version.
             :param vault (object): instance of vault for reading configuration downloader-api.
-
-        Returns:
-            None
-
-        Attributes:
-            :attribute configuration (dict): dictionary with configuration parameters for instagram api communication.
-            :attribute client (object): instance of the instagram api client.
-
-        Examples:
-            >>> configuration = {
-            ...     'username': 'my_username',
-            ...     'password': 'my_password',
-            ...     'session-file': 'data/session.json',
-            ...     'delay-requests': 1,
-            ...     '2fa-enabled': False,
-            ...     '2fa-seed': 'my_seed_secret',
-            ...     'locale': 'en_US',
-            ...     'country-code': '1',
-            ...     'timezone-offset': 10800,
-            ...     'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
-            ...     'proxy-dsn': 'http://localhost:8080'
-            ...     'request-timeout': 10
-            ... }
-            >>> vault = Vault()
-            >>> downloader = Downloader(configuration, vault)
         """
         if not vault:
             raise WrongVaultInstance("Wrong vault instance, you must pass the vault instance to the class argument.")
@@ -85,22 +115,16 @@ class Downloader:
         log.info('[Downloader]: Creating a new instance...')
         self.client = Client()
 
-        log.info('[Downloader]: Configuring client settings...')
+        log.info('[Downloader]: Setting up the client configuration...')
         self.client.delay_range = [1, int(self.configuration['delay-requests'])]
         self.client.request_timeout = int(self.configuration['request-timeout'])
-        self.client.set_locale(locale=self.configuration['locale'])
-        self.client.set_country_code(country_code=int(self.configuration['country-code']))
-        self.client.set_timezone_offset(seconds=int(self.configuration['timezone-offset']))
-        self.client.set_user_agent(user_agent=self.configuration['user-agent'])
         self.client.set_proxy(dsn=self.configuration.get('proxy-dsn', None))
-        log.info('[Downloader]: Client settings: %s', self.client.get_settings())
-
-        auth_status = self.login()
-        if auth_status == 'logged_in':
-            log.info('[Downloader]: Instance created successfully with account %s', self.configuration['username'])
-        else:
-            raise FailedAuthInstagram("Failed to authenticate the Instaloader instance.")
-
+        self.general_settings_list = [
+            'locale', 'country_code', 'country', 'timezone_offset'
+        ]
+        self.device_settings_list = [
+            'app_version', 'version_code', 'manufacturer', 'model', 'device', 'cpu', 'dpi', 'resolution', 'android_release', 'android_version'
+        ]
         self.download_methods = {
             (1, 'any'): self.client.photo_download,
             (2, 'feed'): self.client.video_download,
@@ -108,6 +132,113 @@ class Downloader:
             (2, 'igtv'): self.client.igtv_download,
             (8, 'any'): self.client.album_download
         }
+
+        auth_status = self.login()
+        if auth_status == 'logged_in':
+            log.info('[Downloader]: Instance created successfully with account %s', self.configuration['username'])
+        else:
+            raise FailedAuthInstagram("Failed to authenticate the Instaloader instance.")
+
+    def _get_login_args(self) -> dict:
+        """Get login arguments for the Instagram API"""
+        if self.configuration['2fa-enabled']:
+            totp_code = self.client.totp_generate_code(seed=self.configuration['2fa-seed'])
+            log.info('[Downloader]: Two-factor authentication is enabled. TOTP code: %s', totp_code)
+            return {
+                'username': self.configuration['username'],
+                'password': self.configuration['password'],
+                'verification_code': totp_code
+            }
+        return {
+            'username': self.configuration['username'],
+            'password': self.configuration['password']
+        }
+
+    def _create_new_session(self, login_args: dict) -> None:
+        """Create a new session file for the Instagram API"""
+        self._set_session_settings()
+        self.client.login(**login_args)
+        self.client.dump_settings(self.configuration['session-file'])
+        log.info('[Downloader]: The new session file was created successfully: %s', self.configuration['session-file'])
+
+    def _handle_relogin(self, login_args: dict) -> None:
+        """Handle re-authentication in the Instagram API"""
+        log.info('[Downloader]: Authentication with the clearing of the session...')
+        old_uuids = self.client.get_settings().get("uuids", {})
+        self.client.set_settings({})
+        self.client.set_uuids(old_uuids)
+        self._create_new_session(login_args)
+
+    def _load_session(self, login_args: dict) -> None:
+        """Load or create a session."""
+        log.info('[Downloader]: Authentication with the existing session...')
+        session_file = self.configuration['session-file']
+        if os.path.exists(session_file):
+            self.client.load_settings(session_file)
+            # Temporarily fix for country, because it is not working in set_settings
+            self.client.set_country(country=self.configuration['country'])
+            if not self._validate_session_settings():
+                self._create_new_session(login_args)
+        else:
+            self._create_new_session(login_args)
+
+    def _set_session_settings(self) -> None:
+        """
+        The method for setting general session settings for the Instagram API, such as
+            - locale
+            - country code
+            - country
+            - timezone offset
+            - device settings
+            - user agent
+        """
+        log.info('[Downloader]: Extracting device settings...')
+        device_settings = json.loads(self.configuration['device-settings'])
+        if not all(item in device_settings.keys() for item in self.device_settings_list):
+            raise ValueError("Incorrect device settings in the configuration. Please check the configuration in the Vault.")
+
+        # Extract other settings except device settings
+        log.info('[Downloader]: Extracting other settings...')
+        other_settings = {item: None for item in self.general_settings_list}
+        for item in other_settings.keys():
+            other_settings[item] = self.configuration[item.replace('_', '-')]
+
+        log.debug('[Downloader]: Retrieved settings: %s', {**other_settings, 'device_settings': device_settings})
+
+        # Apply all session settings
+        self.client.set_settings(settings={**other_settings, 'device_settings': device_settings})
+        # Temporarily fix for country
+        # Country in set_settings is not working
+        self.client.set_country(country=other_settings['country'])
+        self.client.set_user_agent()
+        log.info('[Downloader]: General session settings have been successfully set: %s', self.client.get_settings())
+
+    def _validate_session_settings(self) -> bool:
+        """
+        The method for checking the correctness between the session settings and the configuration settings.
+
+        Returns:
+            (bool) True if the session settings are equal to the configuration settings, otherwise False.
+        """
+        log.info('[Downloader]: Checking the difference between the session settings and the configuration settings...')
+        session_settings = self.client.get_settings()
+        for item in self.general_settings_list:
+            if str(session_settings[item]) != str(self.configuration[item.replace('_', '-')]):
+                log.info(
+                    '[Downloader]: The session key value are not equal to the expected value: %s != %s. Session will be reset',
+                    session_settings[item], self.configuration[item.replace('_', '-')]
+                )
+                return False
+        device_settings = self.client.get_settings()['device_settings']
+        for item in self.device_settings_list:
+            if str(device_settings[item]) != str(json.loads(self.configuration['device-settings'])[item]):
+                log.info(
+                    '[Downloader]: The session key value are not equal to the expected value: %s != %s. Session will be reset',
+                    device_settings[item], json.loads(self.configuration['device-settings'])[item]
+                )
+                return False
+        log.info('[Downloader]: The session settings are equal to the expected settings.')
+        return True
 
     @staticmethod
     def exceptions_handler(method) -> None:
@@ -118,18 +249,23 @@ class Downloader:
             :param method (function): method to be wrapped.
         """
         def wrapper(self, *args, **kwargs):
+            random_shift = random.randint(600, 7200)
             try:
                 return method(self, *args, **kwargs)
             except LoginRequired:
-                log.error('[Downloader]: Instagram API login required. Re-authentication...')
+                log.error('[Downloader]: Instagram API login required. Re-authenticate after %s minutes', round(random_shift/60))
+                time.sleep(random_shift)
+                log.info('[Downloader]: Re-authenticate after timeout due to login required')
                 self.login(method='relogin')
             except ChallengeRequired:
-                log.error('[Downloader]: Instagram API requires challenge. Need manually pass in browser. Retry after 1 hour')
-                time.sleep(3600)
+                log.error('[Downloader]: Instagram API requires challenge in browser. Retry after %s minutes', round(random_shift/60))
+                time.sleep(random_shift)
+                log.info('[Downloader]: Re-authenticate after timeout due to challenge required')
                 self.login()
             except PleaseWaitFewMinutes:
-                log.error('[Downloader]: Device or IP address has been restricted. Just wait a one hour and try again')
-                time.sleep(3600)
+                log.error('[Downloader]: Device or IP address has been restricted. Just wait a %s minutes and retry', round(random_shift/60))
+                time.sleep(random_shift)
+                log.info('[Downloader]: Retry after timeout due to restriction')
                 self.login(method='relogin')
             except (ReadTimeoutError, RequestsConnectionError, ClientRequestTimeout):
                 log.error('[Downloader]: Timeout error downloading post content. Retry after 1 minute')
@@ -145,8 +281,8 @@ class Downloader:
         Args:
             :param method (str): the type of authentication in the Instagram API. Default: 'session'.
                 possible values:
-                    'session' - authentication by existing session file. Or create session file for existing device.
-                    'relogin' - authentication as an existing device. This will create a new session file and clear the old attributes.
+                    'session' - authentication by existing session file or create session file for existing device.
+                    'relogin' - authentication as an existing device. Creates a new session file and clears the old attributes.
 
         Returns:
             (str) logged_in
@@ -155,37 +291,14 @@ class Downloader:
         """
         log.info('[Downloader]: Authentication in the Instagram API with type: %s', method)
 
-        # 2FA authentication settings
-        if self.configuration['2fa-enabled']:
-            totp_code = self.client.totp_generate_code(seed=self.configuration['2fa-seed'])
-            log.info('[Downloader]: Two-factor authentication is enabled. TOTP code: %s', totp_code)
-            login_args = {
-                'username': self.configuration['username'],
-                'password': self.configuration['password'],
-                'verification_code': totp_code
-            }
-        else:
-            login_args = {
-                'username': self.configuration['username'],
-                'password': self.configuration['password']
-            }
+        # Generate login arguments
+        login_args = self._get_login_args()
 
-        # Login to the Instagram API
-        if method == 'session' and os.path.exists(self.configuration['session-file']):
-            log.info('[Downloader]: Loading session file with creation date %s', time.ctime(os.path.getctime(self.configuration['session-file'])))
-            self.client.load_settings(self.configuration['session-file'])
-            self.client.login(**login_args)
-        elif method == 'relogin':
-            log.info('[Downloader]: Relogin to the Instagram API...')
-            old_session = self.client.get_settings()
-            self.client.set_settings({})
-            self.client.set_uuids(old_session["uuids"])
-            self.client.login(**login_args)
-            self.client.dump_settings(self.configuration['session-file'])
+        # Handle authentication method
+        if method == 'relogin':
+            self._handle_relogin(login_args)
         else:
-            log.info('[Downloader]: Creating a new session file...')
-            self.client.login(**login_args)
-            self.client.dump_settings(self.configuration['session-file'])
+            self._load_session(login_args)
 
         # Check the status of the authentication
         log.info('[Downloader]: Checking the status of the authentication...')
@@ -252,8 +365,8 @@ class Downloader:
                     'status': status if status else 'failed'
                 }
 
-        except (MediaUnavailable, MediaNotFound) as error:
-            log.warning('[Downloader]: Post %s not found, perhaps it was deleted. Message will be marked as processed: %s', shortcode, error)
+        except (MediaUnavailable, MediaNotFound):
+            log.warning('[Downloader]: Post %s not found, perhaps it was deleted. Message will be marked as processed', shortcode)
             response = {
                 'post': shortcode,
                 'owner': 'undefined',